requirements.txt
fastapi 
uvicorn
gunicorn
langchain
langchain-openai
langchain-core
langchain-community
langchain-experimental
openai
python-dotenv
tiktoken
jupyterlab
flask
pydrive2 
py2neo 
fitz 
pymupdf
google-auth-oauthlib 
google-auth
neo4j
tiktoken
yfiles_jupyter_graphs
faiss-cpu
fuzzywuzzy
python-Levenshtein
pypdf
flask-cors
json-repair
